+++ devel.R
@@ -29,6 +29,3 @@ if (FALSE) {
 
 
 
-
-
-add_github_url_to_desc()